Move integration helpers into own module
