+++ natskell.cabal
@@ -24,7 +24,8 @@ library
         bytestring ^>=0.11.3.0,
         word8 ^>=0.1.3,
         aeson ^>=2.0.3.0,
-        text ^>= 1.2.3.1
+        text ^>= 1.2.3.1,
+        network-simple ^>= 0.4.5
 
     -- Modules included in this library but not exported.
     exposed-modules:
@@ -39,6 +40,7 @@ library
         Connect
         Sub
         Unsub
+        Client
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
@@ -71,4 +73,6 @@ test-suite natskell-test
         ConnectSpec
         SubSpec
         UnsubSpec
+        Integration
+        ClientSpec
